+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 # Changelog
 
 ### Unreleased
-* Added support for `roundTo` field in availability response
-* 
+* Added support for `roundTo` field in availability response model
+* Added support for `attributes` field in folder model
 
 ### 7.1.0 / 2024-02-12
 * Added support for `/v3/connect/tokeninfo` endpoint