+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-### Unreleased
+### 7.2.0 / 2024-02-27
 * Added support for `roundTo` field in availability response model
 * Added support for `attributes` field in folder model
 * Added support for icloud as an auth provider